Add Sci-hub-Fy to context menu. Fix #1.
we are adding two kind of actions: "page" and "link". If user right-
click a link and select Sci-Hub-Fy, we append ".sci-hub.io" to the link
and redirect the user to it. If user right-click anywhere but a link in
a page, we append ".sci-hub.io" to current tab URL and redirect user to
it.